\name{analyzeRiksdag-package}
\alias{analyzeRiksdag-package}
\alias{analyzeRiksdag}
\docType{package}
\title{
\packageTitle{analyzeRiksdag}
}
\description{
\packageDescription{analyzeRiksdag}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{analyzeRiksdag}
\packageIndices{analyzeRiksdag}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{analyzeRiksdag}

Maintainer: \packageMaintainer{analyzeRiksdag}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
