% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{fun}
\alias{fun}
\alias{fun_fetch_data}
\alias{fun_get_county_options}
\alias{fun_get_assembly_year_options}
\alias{fun_get_party_options}
\alias{fun_get_counts}
\alias{fun_bar_chart}
\alias{fun_initialize_utskott_csv}
\alias{fun_check_utskott_file}
\alias{get_Riksdag}
\alias{get_utskott}
\alias{get_titlar}
\title{Functions to access and manipulate data from Riksdag API}
\usage{
fun_fetch_data(assembly_year, beteckning, agenda_item)

fun_get_county_options()

fun_get_assembly_year_options()

fun_get_party_options()

fun_get_counts(df, filter)

fun_bar_chart(df, filter)

fun_initialize_utskott_csv()

fun_check_utskott_file()

get_Riksdag()

get_utskott(year)

get_titlar(year, utskott)
}
\description{
Functions to access and manipulate data from Riksdag API
}
\details{
\code{fun_fetch_data()} fetches all data from the Riksdag API using assembly_year, beteckning and agenda_item

\code{fun_get_county_options()} scrapes the options for valkrets from the Riksdag API form

\code{fun_get_assembly_year_options()} scrapes the options for assembly year from the Riksdag API form

\code{fun_get_party_options()} scrapes the options for party from the Riksdag API form

\code{fun_get_counts()} takes the data retrieved by \code{fun_fetch_data()} and gets a count per type of vote based on a filter

\code{fun_bar_chart()} produces a bar chart of vote counts from data retrieved by \code{fun_fetch_data()} and a Gender, Region, or Party filter

\code{fun_initialize_utskott_csv()} scrapes the utskott options from the Riksdag website, adds a hardcoded short alias, and writes them to a csv

\code{fun_check_utskott_file()} checks the length of utskott.csv versus the utskott on the Riksdag website and throws a warning if there is a discrepancy

\code{get_Riksdag()} fetches data from the Riksdag API for every available year and creates a large data frame that can be used to retrieve all available utskott and beteckning for each year

\code{get_titlar()} read in titleframe.rds, filters on year and returns a named list of all available utskott

\code{get_titlar()} reads in titleframe.rds, filters on year and utskott and returns a named list of all available beteckning titles and codes
}
\examples{
fun_fetch_data("2022\%2F23", "AU10", "2")
fun_get_counts(df, "Gender")
fun_bar_chart(df, "Gender")
get_utskott("2022\%2F23")

get_titlar("2022\%2F23", "AU")

}
